Move nonArrivals to booking client
